+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/EnchantmentsFeature.java
@@ -146,6 +146,7 @@ public class EnchantmentsFeature extends JsonFeature {
 			IdTagMatcher.newId("minecraft:looting"),
 			IdTagMatcher.newId("minecraft:fortune"),
 			IdTagMatcher.newId("minecraft:luck_of_the_sea"),
+			IdTagMatcher.newId("minecraft:sweeping"),
 			IdTagMatcher.newId("farmersdelight:backstabbing")
 
 	));